+++ app/controllers/integral/backend/static_pages_controller.rb
@@ -6,6 +6,10 @@ module Integral
       # Dashboard to show website stats and other useful information
       def dashboard; end
 
+      def resource_klass
+        nil
+      end
+
       private
 
       def set_breadcrumbs; end