+++ app/decorators/integral/category_version_decorator.rb
@@ -1,0 +1,7 @@
+module Integral
+  class CategoryVersionDecorator < VersionDecorator
+    def item_icon
+      'tags'
+    end
+  end
+end