+++ app/decorators/integral/image_version_decorator.rb
@@ -1,0 +1,7 @@
+module Integral
+  class ImageVersionDecorator < VersionDecorator
+    def item_icon
+      'image'
+    end
+  end
+end