+++ app/decorators/integral/post_version_decorator.rb
@@ -1,0 +1,7 @@
+module Integral
+  class PostVersionDecorator < VersionDecorator
+    def item_icon
+      'rss'
+    end
+  end
+end