+++ app/helpers/integral/backend/base_helper.rb
@@ -5,6 +5,37 @@ module Integral
     module BaseHelper
       include Integral::SupportHelper
 
+      # @return [String] Integral card
+      def render_card(partial, locals = {})
+        render(partial: "integral/backend/shared/cards/#{partial}", locals: locals)
+      end
+
+      # @return [Array] returns array of VersionDecorators subclassed depending on the Version subclass
+      def recent_user_activity_grid
+        @recent_user_activity_grid ||= begin
+                                      options = { user: current_user.id }
+                                      options[:object] = resource_klass.to_s if resource_klass.present?
+                                      options[:item_id] = @resource.id if @resource.present?
+
+                                      recent_activity_grid(options)
+                                    end
+      end
+
+      # @return [Array] returns array of VersionDecorators subclassed depending on the Version subclass
+      def recent_site_activity_grid
+        @recent_site_activity_grid ||= begin
+                                      options = {}
+                                      options[:object] = resource_klass.to_s if resource_klass.present?
+                                      options[:item_id] = @resource.id if @resource.present?
+
+                                      recent_activity_grid(options)
+                                    end
+      end
+
+      def recent_activity_grid(options)
+        Integral::Grids::ActivitiesGrid.new(options)
+      end
+
       # @return [String] title provided through yield or i18n scoped to controller namespace & action
       def page_title
         return content_for(:title) if content_for?(:title)
@@ -59,6 +90,15 @@ module Integral
         [
           { scope: Integral::Post.published, label: 'Published' },
           { scope: Integral::Post.draft, label: 'Draft ' }
+        ]
+      end
+
+      # Donut Graph - At a Glance
+      def dataset_at_a_glance_pages
+        [
+          { scope: Integral::Page.published, label: 'Published' },
+          { scope: Integral::Page.draft, label: 'Draft ' },
+          { scope: Integral::Page.archived, label: 'Archived ' }
         ]
       end