+++ db/migrate/20200401210442_create_category_versions.rb
@@ -1,0 +1,20 @@
+class CreateCategoryVersions < ActiveRecord::Migration[5.2]
+  # The largest text column available in all supported RDBMS is
+  # 1024^3 - 1 bytes, roughly one gibibyte.  We specify a size
+  # so that MySQL will use `longtext` instead of `text`.  Otherwise,
+  # when serializing very large objects, `text` might not be big enough.
+  TEXT_BYTES = 1_073_741_823
+
+  def change
+    create_table :integral_category_versions do |t|
+      t.string   :item_type, {:null=>false}
+      t.integer  :item_id,   null: false
+      t.string   :event,     null: false
+      t.string   :whodunnit
+      t.text     :object, limit: TEXT_BYTES
+      t.text     :object_changes, limit: TEXT_BYTES
+      t.datetime :created_at
+    end
+    add_index :integral_category_versions, %i(item_type item_id)
+  end
+end