+++ spec/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_12_03_090008) do
+ActiveRecord::Schema.define(version: 2020_04_01_210442) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -50,6 +50,17 @@ ActiveRecord::Schema.define(version: 2019_12_03_090008) do
     t.datetime "updated_at", null: false
     t.bigint "image_id"
     t.index ["image_id"], name: "index_integral_categories_on_image_id"
+  end
+
+  create_table "integral_category_versions", force: :cascade do |t|
+    t.string "item_type", null: false
+    t.integer "item_id", null: false
+    t.string "event", null: false
+    t.string "whodunnit"
+    t.text "object"
+    t.text "object_changes"
+    t.datetime "created_at"
+    t.index ["item_type", "item_id"], name: "index_integral_category_versions_on_item_type_and_item_id"
   end
 
   create_table "integral_enquiries", id: :serial, force: :cascade do |t|