+++ spec/rails_helper.rb
@@ -4,7 +4,7 @@ ENV['RAILS_ENV'] ||= 'test'
 # Analysis code coverage
 require 'simplecov'
 SimpleCov.start 'rails'
-SimpleCov.minimum_coverage 85.95
+SimpleCov.minimum_coverage 88.10
 
 require File.expand_path('../dummy/config/environment', __FILE__)
 # Prevent database truncation if the environment is production
@@ -24,6 +24,7 @@ require 'capybara/poltergeist'
 require 'paper_trail/frameworks/rspec'
 
 Capybara.default_driver = :poltergeist
+Capybara.server = :webrick
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
@@ -64,6 +65,7 @@ RSpec.configure do |config|
 
   # Add feature helpers
   config.include Features::Helpers, type: :feature
+  config.include Features::Helpers, type: :system
 
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"