+++ spec/spec_helper.rb
@@ -105,6 +105,10 @@ RSpec.configure do |config|
     Rails.application.load_seed
   end
 
+  config.before(:each, type: :system) do
+    driven_by :poltergeist
+  end
+
   config.around :each do |ex|
     ex.run_with_retry retry: 3
   end